First try at vectorization + gitignore
First, I added a gitignore so that most things wont be added. I kept the timing.pdf, so that we can see results from others as we add things.

Second, I edited dgemm_mine.c to loop through multiple entries of C at the same time. I think it's in a good position to be combined with alignment now. It's a bit faster!